fix: decople inventory routes from branch, and improve structure
